refactor vite-config, add reset css
