+++ .env
@@ -1,1 +1,2 @@
 VITE_ADD_RED_FRAME=0
+
