+++ index.html
@@ -1,6 +1,7 @@
 <!DOCTYPE html>
 <html lang="en">
 <head>
+  <link rel="stylesheet" href="/reset_style.css">
   <script  defer src="//cdn.jsdelivr.net/npm/prebid.js@latest/dist/not-for-prod/prebid.js"></script>
   <script type="module" src="/main.js" defer></script>
   <title>Ad</title>
@@ -16,7 +17,7 @@
 
 </div>
 <script>
-  wrapper.auctionForPlacement('div-1')
+  // wrapper.auctionForPlacement('div-1')
 </script>
 <br>
